// src/app/common/interfaces/interfaces.ts
export interface CardInfo {
    title: string;
    subtitle?: string;
    description: string;
    logo?: string;
    dates?: string;
    website?: string;
    expandedText?: string[];
    image?: string;
    technologies?: string[];
}

export interface Logo {
    description: string;
    url: string;
}